requirements.txt
attrs==20.3.0
PyGuitarPro==0.8
